Revert "Merge branch 'master' of https://github.com/ReventonesInc/Proyecto-ProCable"
This reverts commit 3c20a9fee512504f37f1e3c8a1d9134a146f65cc, reversing
changes made to f0e2cef7405e0c7c4c1d8be3b27d4bfb6249e153.

+++ ProCableWW1/src/ventanas/VentanaAgregarPlan.java
@@ -162,7 +162,7 @@ public class VentanaAgregarPlan extends javax.swing.JFrame {
                     JOptionPane.showMessageDialog(null, "Es obligatorio completar el formulario!");
                 }
                 else{
-                    float valor = Float.parseFloat(jTextValor.getText());
+                    int valor = Integer.parseInt(jTextValor.getText());
                     int cantidadCanal = Integer.parseInt(jTextCanales.getText());
                     if((this.procable).agregarPlan(jTextNombre.getText(), valor, cantidadCanal) == true){
                         JOptionPane.showMessageDialog(null, "Plan agregado con exito!");
